Store config data on session struct
This makes it easier to start a session and end it without
having to keep track of an additional config value. The session
id isn't really useful without the base_url in the config of
which WebDriver server to connect to. This is an attempt at
making bad states not possible.